Wrote first draft of our explanation for how Deleting/Archiving works
Some resources cannot be deleted in certain cases (instead they can
only be archived), due to the fact that there is data associated with
this resource which depends on this resource existing.

We also cannot archive these resources if they can be deleted,
archiving is an action reserved only for resources that have dependent
data. Therefore, we have two separate actions (each with explanations
in their error messages which indicate why the action cannot be
performed).

This is the first draft of our explanation for how this system works.