User Delete and Deactivate Actions now follow Delete/Archive pattern
The User Delete and Deactivate Actions are the same as the
Delete/Archive actions we have previously established, but we use a
different verb for Archiving when discussing users. To keep things
consistent between the application and the API, we have updated our
specification to indicate that these actions are the same, but use
different names.