Allow for multiple entries to be marked as invoiced outside of Freckle
To help make the API easier to use for consumers (particularly
consumers who previously or currently use a different invoicing system),
We have added support for multiple entries to be marked as invoiced
outside of Freckle.

This will also make it easier for external services that include invoicing to integrate with Freckle, since they don't have to make
multiple requests to the API in order to mark entries as invoiced.