Redesigned default layout to have sidebar navigation and Logo header
This is the first step towards a cleaner design for the Freckle API.

I have changed the site's layout so that we use as window space as
possible (instead of having a set width on our content). I'm using
Flexbox to allow us to quickly create a sidebar for our navigation
to each resource.

The Freckle Logo is now at the top, which will also contain links to
to our help site, the app's status page, and other important locations